Replacing dot_expression with _dot_or_index_chain
This mimics more closely how the compiler sees this. Especially with the
recent changes announced, to move; 3+ length do_expressions will occur.

The code prior to this change would treat the dot_expression as the root
of the subtree.  This would lead to issues if the expression did not
have an expression_term in that slot.

This new version puts the call receiving at the root.  This is also
necessary for Semgrep grammer to work correclty.